chore: Move onInitializeOvermind & internalAction into global namespace
