+++ public/src/overmind/namespaces/global/internalActions.ts
@@ -1,7 +1,7 @@
 import { Code } from "@connectrpc/connect"
-import { Context } from "."
-import { RepositoryRequest } from "../../proto/qf/requests_pb"
-import { Prompt, promptOnErrorResponse } from "./utils/errors"
+import { Context } from "../.."
+import { RepositoryRequest } from "../../../../proto/qf/requests_pb"
+import { Prompt, promptOnErrorResponse } from "../../utils/errors"
 
 export const isEmptyRepo = async (
   { effects }: Context,
